requirements.txt
alabaster==0.7.12
alembic==1.0.0
arrow==1.2.2
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.7.0
backcall==0.2.0
Beaker==1.11.0
beautifulsoup4==4.9.1
binaryornot==0.4.4
bleach==3.1.4
blinker==1.4
certifi==2021.5.30
chardet==4.0.0
-e git+https://github.com/ckan/ckan.git@65af260bef26ea99981597f872d39da0a3f4f3f9#egg=ckan
ckanapi==4.6
# Editable install with no version control (ckanext-metadata-form==0.0.1)
-e /home/mohab/Main/development/WRO/src/ckanext-metadata_form
-e git+https://github.com/ckan/ckanext-scheming.git@e24ebca2e29cad8c8061755e55373906679506a8#egg=ckanext_scheming
ckantoolkit==0.0.4
click==7.1.2
cookiecutter==1.7.0
coverage==6.3.2
coveralls==3.3.1
decorator==5.1.1
docopt==0.6.2
docutils==0.12
dominate==2.4.0
executing==0.8.3
factory-boy==2.12.0
Faker==13.3.4
fanstatic==1.1
feedgen==0.9.0
Flask==1.1.1
Flask-Babel==1.0.0
Flask-DebugToolbar==0.11.0
flask-multistatic==1.0
freezegun==0.3.15
funcsigs==1.0.2
future==0.18.2
idna==2.10
imagesize==1.3.0
importlib-metadata==4.11.3
incremental==21.3.0
ipdb==0.13.2
ipython==8.2.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.10.1
jinja2-time==0.2.0
lxml==4.6.3
Mako==1.1.5
Markdown==2.6.7
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
mock==2.0.0
more-itertools==8.12.0
nose==1.3.7
packaging==21.3
parso==0.8.3
passlib==1.6.5
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
pip-tools==5.1.2
pluggy==0.13.1
polib==1.0.7
poyo==0.5.0
prompt-toolkit==3.0.29
psycopg2==2.8.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.5.0
pyfakefs==3.2
Pygments==2.11.2
PyJWT==1.7.1
pyparsing==3.0.7
pysolr==3.6.0
pytest==4.6.5
pytest-cov==2.7.1
pytest-freezegun==0.4.1
pytest-rerunfailures==8.0
pytest-split-tests==1.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-magic==0.4.15
python-slugify==6.1.1
pytz==2016.7
PyUtilib==5.7.1
PyYAML==5.4.1
redis==3.5.3
repoze.lru==0.7
repoze.who==2.3
requests==2.25.1
responses==0.10.14
Routes==1.13
rq==1.0
shutilwhich==1.1.0
simplejson==3.10.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==1.8.5
sphinx-rtd-theme==0.4.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
SQLAlchemy==1.3.5
sqlparse==0.2.2
stack-data==0.2.0
text-unidecode==1.3
toml==0.10.2
towncrier==19.2.0
traitlets==5.1.1
tzlocal==1.3
unicodecsv==0.14.1
urllib3==1.26.6
uWSGI==2.0.20
watchdog==2.1.5
wcwidth==0.2.5
webassets==0.12.1
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==1.0.0
whichcraft==0.6.1
zipp==3.8.0
zope.interface==4.3.2
